rename submit notes
