+++ frontend/src/js/components/MentorDashboard/NoteList.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+export default (props) => {
+  return (
+    <div> {props.menteenotes.map( (note) => {
+      return (<div>
+
+      </div>
+    )
+    }
+
+    )}
+    </div>
+  )
+}
